refactor: simplify chart HTML and extract SVGs
We've added some new HTML here, but this was exported from Figma. And so there's a lot of unnecessary classes and SVGs that are hard-coded in here. So let's do two things. Let's move the SVGs into their own separate files and reference the SVG images. Reference the SVGs as image tags. And then the next thing is remove all of like 95% of the classes, almost all of the classes. Just get rid of them. We don't really need all those classes. Just keep the bare minimal classes because this is just for mobile. There's also a responsive desktop version that we've got to accommodate for, and this just locks in a lot of stuff. So just preserve some of the content and remove most of the classes. Keep critical, critical classes, and then put the SVGs as images. Yes.

```git-revs
cddd738  (Base revision)
5663c5e  Create SVG file for arrow icon
e5e6b64  Create SVG file for vertical green line
a829065  Create SVG file for shorter vertical green line
4ecb6ca  Create SVG file for curved arrow
5bf3c47  Snapshot before codemcp change
HEAD     Refactor the new-charts-text section to simplify HTML and use image tags for SVGs
```

codemcp-id: 23-refactor-simplify-chart-html-and-extract-svgs